refactor:main以外のカスタムフック作成
