+++ hooks/useSignup.ts
@@ -1,0 +1,45 @@
+import React, { useState } from 'react';
+import { useRouter } from 'next/navigation';
+
+import { signup } from '../features/api/auth';
+
+const useSignup = () => {
+  const [name, setName] = useState<string>('');
+  const [email, setEmail] = useState<string>('');
+  const [password, setPassword] = useState<string>('');
+  const [passwordConfirmation, setPasswordConfirmation] = useState<string>('');
+  const [error, setError] = useState<string>('');
+
+  const router = useRouter();
+
+  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    try {
+      const user = await signup(name, email, password, passwordConfirmation);
+      console.log('Signed up user:', user);
+      router.push('/login');
+    } catch (err: unknown) {
+      if (err instanceof Error) {
+        setError(err.message);
+      } else {
+        setError('An unexpected error occurred');
+      }
+    }
+  };
+
+  return {
+    setName,
+    email,
+    setEmail,
+    password,
+    setPassword,
+    passwordConfirmation,
+    setPasswordConfirmation,
+    error,
+    handleSubmit
+  };
+}
+
+export default useSignup;
+
+
